Start setting up docs, rename key_string and cert_string to key and cert, start on examples, split apart files
