+++ csp_adapter_symphony/__init__.py
@@ -2,3 +2,5 @@ __version__ = "0.2.1"
 
 from .adapter import *
 from .adapter_config import *
+from .mention import *
+from .message import *
