+++ docs/wiki/Examples.md
@@ -1,0 +1,10 @@
+The below examples are available [in-source](https://github.com/Point72/csp-adapter-symphony/blob/main/csp_adapter_symphony/examples/).
+They assume the presence of `.key` and `.cert` files in the run directory.
+You will need to modify them to point at your Symphony URL.
+
+# Hello!
+
+Here is a simple example that replies "Hello!" someone says `hello` in a room or direct message to the bot.
+It is available in-source at [`csp_adapter_symphony/examples/hello.py`](https://github.com/Point72/csp-adapter-symphony/blob/main/csp_adapter_symphony/examples/hello.py).
+
+Coming Soon!
