--- docs/wiki/First-Steps.md
@@ -1,1 +1,0 @@
-Coming Soon!
