--- docs/wiki/Home.md
@@ -1,19 +1,0 @@
-# csp-adapter-symphony
-
-A [csp](https://github.com/point72/csp) adapter for [symphony](https://symphony.com)
-
-## Get Started
-
-- [Learn more about csp](https://github.com/Point72/csp)
-- [Install csp-adapter-symphony](Installation) and [get started](First-Steps)
-
-> [!TIP]
-> Find relevant docs with GitHub’s search function, use `repo:Point72/csp-adapter-symphony type:wiki <search terms>` to search the documentation Wiki Pages.
-
-## Community
-
-- [Contribute](Contribute) to `csp-adapter-symphony` and help improve the project
-
-## License
-
-`csp-adapter-symphony` is licensed under the Apache 2.0 license. See the [LICENSE](https://github.com/Point72/csp-adapter-symphony/blob/main/LICENSE) file for details.
